refactor: NBT-65 userId 타입변환

+++ src/main/java/com/newbit/user/repository/UserRepository.java
@@ -5,12 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUserNameAndPhoneNumber(String userName, String phoneNumber);
     Optional<User> findByEmail(String email);
 
-    default User getById(Integer userId){
+    default User getById(Long userId){
         return findById(userId).orElseThrow();
     }
